refactor(gateway-cli): move the gateway auth and login tests to sibling files
Pure move, no behavior change. `src/auth/gateway/auth.rs` and `login.rs` had
grown well past AGENTS.md's "preferably under 500 lines", almost entirely on
their inline test modules. Both now use `#[cfg(test)] mod tests;` pointing at a
sibling file, matching the convention already used by 13 files under `src/`,
`src/auth/slots.rs` among them.

Nothing but the module boundary changed: no production code was relocated, no
signature altered, and no visibility widened — the parent files gain exactly two
lines each and lose the rest. The tests keep their existing `use super::*;`,
which resolves the same way from the sibling module.

  auth.rs   1208 -> 534, plus auth/tests.rs   673
  login.rs  1013 -> 456, plus login/tests.rs  559

`auth.rs` lands at 534, still slightly over the guideline. Production code is
deliberately not split to close that gap: the remaining 534 lines are one
subsystem — discovery, refresh, expiry, and token resolution share the session
and the module doc's body-before-status rule — and cutting them into two files
to satisfy a line count would trade a focused module for two unclear ones. The
guideline says "preferably", and this is the case it leaves room for.

The two test files are over 500 as well, for the same reason: a test module for
one subsystem is focused by construction, and splitting it would separate tests
from the thing they describe.

Test count before and after: 2012 passed, identical — the check that catches a
silently dropped `#[cfg(test)]` module or test function, which is the failure
mode a move like this is vulnerable to.

Part of #353.
